+++ lib/src/terminal/src/ui/gesture/gesture_detector.dart
@@ -10,7 +10,6 @@ class TerminalGestureDetector extends StatefulWidget {
     super.key,
     this.child,
     this.onSingleTapUp,
-    this.onTapUp,
     this.onTapDown,
     this.onSecondaryTapDown,
     this.onSecondaryTapUp,
@@ -25,8 +24,6 @@ class TerminalGestureDetector extends StatefulWidget {
   });
 
   final Widget? child;
-
-  final GestureTapUpCallback? onTapUp;
 
   final GestureTapUpCallback? onSingleTapUp;