+++ lib/src/terminal/src/ui/gesture/gesture_handler.dart
@@ -16,7 +16,6 @@ class TerminalGestureHandler extends StatefulWidget {
     required this.terminalView,
     required this.terminalController,
     this.child,
-    this.onTapUp,
     this.onSingleTapUp,
     this.onTapDown,
     this.onSecondaryTapDown,
@@ -31,8 +30,6 @@ class TerminalGestureHandler extends StatefulWidget {
   final TerminalController terminalController;
 
   final Widget? child;
-
-  final GestureTapUpCallback? onTapUp;
 
   final GestureTapUpCallback? onSingleTapUp;
 
@@ -64,7 +61,6 @@ class _TerminalGestureHandlerState extends State<TerminalGestureHandler> {
   @override
   Widget build(BuildContext context) {
     return TerminalGestureDetector(
-      onTapUp: widget.onTapUp,
       onSingleTapUp: onSingleTapUp,
       onTapDown: onTapDown,
       onSecondaryTapDown: onSecondaryTapDown,